+++ library/rsa_internal.c
@@ -24,7 +24,7 @@
 
 #include "mbedtls/rsa.h"
 #include "mbedtls/bignum.h"
-#include "mbedtls/rsa_internal.h"
+#include "rsa_internal.h"
 
 /*
  * Compute RSA prime factors from public and private exponents
